pyproject.toml
[tool.poetry]
name = "iskra"
version = "0.1.0"
description = ""
authors = ["Romeo Disca <romeo.disca@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.1"
pyserial = "^3.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
